fixed requirements to working versions and added download scripts for windows in runGan

+++ runGan.py
@@ -8,7 +8,7 @@ runcase == 3    training TecoGAN
 runcase == 4    training FRVSR
 runcase == ...  coming... data preparation and so on...
 '''
-import os, subprocess, sys, datetime, signal, shutil
+import os, subprocess, sys, datetime, signal, shutil, platform
 
 runcase = int(sys.argv[1])
 print ("Testing test case %d" % runcase)
@@ -38,31 +38,61 @@ def folder_check(path):
     
     return path
 
-if( runcase == 0 ): # download inference data, trained models
-    # download the trained model
-    if(not os.path.exists("./model/")): os.mkdir("./model/")
-    cmd1 = "wget https://ge.in.tum.de/download/data/TecoGAN/model.zip -O model/model.zip;"
-    cmd1 += "unzip model/model.zip -d model; rm model/model.zip"
-    subprocess.call(cmd1, shell=True)
-    
-    # download some test data
-    cmd2 = "wget https://ge.in.tum.de/download/data/TecoGAN/vid3_LR.zip -O LR/vid3.zip;"
-    cmd2 += "unzip LR/vid3.zip -d LR; rm LR/vid3.zip"
-    subprocess.call(cmd2, shell=True)
-    
-    cmd2 = "wget https://ge.in.tum.de/download/data/TecoGAN/tos_LR.zip -O LR/tos.zip;"
-    cmd2 += "unzip LR/tos.zip -d LR; rm LR/tos.zip"
-    subprocess.call(cmd2, shell=True)
-    
-    # download the ground-truth data
-    if(not os.path.exists("./HR/")): os.mkdir("./HR/")
-    cmd3 = "wget https://ge.in.tum.de/download/data/TecoGAN/vid4_HR.zip -O HR/vid4.zip;"
-    cmd3 += "unzip HR/vid4.zip -d HR; rm HR/vid4.zip"
-    subprocess.call(cmd3, shell=True)
-    
-    cmd3 = "wget https://ge.in.tum.de/download/data/TecoGAN/tos_HR.zip -O HR/tos.zip;"
-    cmd3 += "unzip HR/tos.zip -d HR; rm HR/tos.zip"
-    subprocess.call(cmd3, shell=True)
+if( runcase == 0 ): # download inference data, trained models#
+
+    if platform.system() == 'Windows':
+        # download the trained model
+        if(not os.path.exists("./model/")): os.mkdir("./model/")
+        cmd1 = "curl.exe --output .\model\model.zip --url https://ge.in.tum.de/download/data/TecoGAN/model.zip &"
+        cmd1 += "powershell.exe -NoP -NonI -Command \"Expand-Archive './model/model.zip' 'model'\" & del .\model\model.zip"
+        subprocess.call(cmd1, shell=True)
+        
+        # download some test data
+        if(not os.path.exists("./LR/")): os.mkdir("./LR/")
+        cmd2 = "curl.exe --output .\LR\\vid3.zip --url https://ge.in.tum.de/download/data/TecoGAN/vid3_LR.zip &"
+        cmd2 += "powershell.exe -NoP -NonI -Command \"Expand-Archive './LR/vid3.zip' 'LR'\" & del .\LR\\vid3.zip"
+        subprocess.call(cmd2, shell=True)
+        
+        cmd2 = "curl.exe --output .\LR\\tos.zip --url https://ge.in.tum.de/download/data/TecoGAN/tos_LR.zip &"
+        cmd2 += "powershell.exe -NoP -NonI -Command \"Expand-Archive './LR/tos.zip' 'LR'\" & del .\LR\\tos.zip"
+        subprocess.call(cmd2, shell=True)
+        
+        # download the ground-truth data
+        if(not os.path.exists("./HR/")): os.mkdir("./HR/")
+        cmd3 = "curl.exe --output .\HR\\vid4.zip --url https://ge.in.tum.de/download/data/TecoGAN/vid4_HR.zip &"
+        cmd3 += "powershell.exe -NoP -NonI -Command \"Expand-Archive '.\HR\\vid4.zip' 'HR'\" & del .\HR\\vid4.zip"
+        subprocess.call(cmd3, shell=True)
+        
+        cmd3 = "curl.exe --output .\HR\\tos.zip --url https://ge.in.tum.de/download/data/TecoGAN/tos_HR.zip &"
+        cmd3 += "powershell.exe -NoP -NonI -Command \"Expand-Archive '.\HR\\tos.zip' 'HR'\" & del .\HR\\tos.zip"
+        subprocess.call(cmd3, shell=True)
+
+    elif platform.system() == 'Linux':
+        # download the trained model
+        if(not os.path.exists("./model/")): os.mkdir("./model/")
+        cmd1 = "wget https://ge.in.tum.de/download/data/TecoGAN/model.zip -O model/model.zip;"
+        cmd1 += "unzip model/model.zip -d model; rm model/model.zip"
+        subprocess.call(cmd1, shell=True)
+
+        # download some test data
+        if(not os.path.exists("./LR/")): os.mkdir("./LR/")
+        cmd2 = "wget https://ge.in.tum.de/download/data/TecoGAN/vid3_LR.zip -O LR/vid3.zip;"
+        cmd2 += "unzip LR/vid3.zip -d LR; rm LR/vid3.zip"
+        subprocess.call(cmd2, shell=True)
+
+        cmd2 = "wget https://ge.in.tum.de/download/data/TecoGAN/tos_LR.zip -O LR/tos.zip;"
+        cmd2 += "unzip LR/tos.zip -d LR; rm LR/tos.zip"
+        subprocess.call(cmd2, shell=True)
+
+        # download the ground-truth data
+        if(not os.path.exists("./HR/")): os.mkdir("./HR/")
+        cmd3 = "wget https://ge.in.tum.de/download/data/TecoGAN/vid4_HR.zip -O HR/vid4.zip;"
+        cmd3 += "unzip HR/vid4.zip -d HR; rm HR/vid4.zip"
+        subprocess.call(cmd3, shell=True)
+
+        cmd3 = "wget https://ge.in.tum.de/download/data/TecoGAN/tos_HR.zip -O HR/tos.zip;"
+        cmd3 += "unzip HR/tos.zip -d HR; rm HR/tos.zip"
+        subprocess.call(cmd3, shell=True)
     
 elif( runcase == 1 ): # inference a trained model
     
@@ -73,7 +103,7 @@ elif( runcase == 1 ): # inference a trained model
     
     # run these test cases one by one:
     for nn in range(len(testpre)):
-        cmd1 = ["python3", "main.py",
+        cmd1 = ["python", "main.py",
             "--cudaID", "0",            # set the cudaID here to use only one GPU
             "--output_dir",  dirstr,    # Set the place to put the results.
             "--summary_dir", os.path.join(dirstr, 'log/'), # Set the place to put the log. 
@@ -97,7 +127,7 @@ elif( runcase == 2 ): # calculate all metrics, and save the csv files, should us
 
     tar_list = [(tarstr+_) for _ in testpre]
     out_list = [(dirstr+_) for _ in testpre]
-    cmd1 = ["python3", "metrics.py",
+    cmd1 = ["python", "metrics.py",
         "--output", dirstr+"metric_log/",
         "--results", ",".join(out_list),
         "--targets", ",".join(tar_list),
@@ -139,7 +169,7 @@ elif( runcase == 3 ): # Train TecoGAN
     now_str = datetime.datetime.now().strftime("%m-%d-%H")
     train_dir = folder_check("ex_TecoGAN%s/"%now_str)
     # train TecoGAN, loss = l2 + VGG54 loss + A spatio-temporal Discriminator
-    cmd1 = ["python3", "main.py",
+    cmd1 = ["python", "main.py",
         "--cudaID", "0", # set the cudaID here to use only one GPU
         "--output_dir", train_dir, # Set the place to save the models.
         "--summary_dir", os.path.join(train_dir,"log/"), # Set the place to save the log. 
@@ -247,7 +277,7 @@ elif( runcase == 3 ): # Train TecoGAN
 elif( runcase == 4 ): # Train FRVSR, loss = l2 warp + l2 content
     now_str = datetime.datetime.now().strftime("%m-%d-%H")
     train_dir = folder_check("ex_FRVSR%s/"%now_str)
-    cmd1 = ["python3", "main.py",
+    cmd1 = ["python", "main.py",
         "--cudaID", "0", # set the cudaID here to use only one GPU
         "--output_dir", train_dir, # Set the place to save the models.
         "--summary_dir", os.path.join(train_dir,"log/"), # Set the place to save the log.
